+++ syncbot/handlers/groups.py
@@ -131,6 +131,7 @@ def handle_create_group(
         callback_id=actions.CONFIG_CREATE_GROUP_SUBMIT,
         title_text="Create Group",
         submit_button_text="Create Group",
+        body=body,
     )
 
 
@@ -229,6 +230,7 @@ def handle_join_group(
         callback_id=actions.CONFIG_JOIN_GROUP_SUBMIT,
         title_text="Join Group",
         new_or_add="new",
+        body=body,
     )
 
 
@@ -416,6 +418,7 @@ def handle_invite_workspace(
             title_text="Oops!",
             submit_button_text=None,
             new_or_add="new",
+            body=body,
         )
         return
 
@@ -475,6 +478,7 @@ def handle_invite_workspace(
         submit_button_text=submit_text,
         parent_metadata={"group_id": group_id},
         new_or_add="new",
+        body=body,
     )